mirror/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
pyyaml==6.0
jsonschema==4.19.0
